+++ llm_guard/output_scanners/factual_consistency.py
@@ -9,22 +9,22 @@ torch = lazy_load_dep("torch")
 _model_path = MODEL_BASE
 
 
-class Refutation(Scanner):
+class FactualConsistency(Scanner):
     """
     Refutation Class:
 
     This class checks for refutation between a given prompt and output using a pretrained NLI model.
     """
 
-    def __init__(self, threshold=0.5):
+    def __init__(self, minimum_score=0.5):
         """
         Initializes an instance of the Refutation class.
 
         Parameters:
-            threshold (float): The threshold used to determine refutation. Defaults to 0.
+            minimum_score (float): The minimum entailment score for the output to be considered valid. Defaults to 0.5.
         """
 
-        self._threshold = threshold
+        self._minimum_score = minimum_score
 
         transformers = lazy_load_dep("transformers")
         self._tokenizer = get_tokenizer(_model_path)
@@ -39,7 +39,7 @@ class Refutation(Scanner):
             return output, True, 0.0
 
         tokenized_input_seq_pair = self._tokenizer(
-            output, prompt, truncation=True, return_tensors="pt"
+            output, prompt, padding=True, truncation=True, return_tensors="pt"
         ).to(device())
         model_output = self._model(tokenized_input_seq_pair["input_ids"])
         model_prediction = torch.softmax(model_output["logits"][0], -1).tolist()
@@ -48,12 +48,12 @@ class Refutation(Scanner):
             name: round(float(pred), 2) for pred, name in zip(model_prediction, label_names)
         }
 
-        not_entailment_score = prediction["not_entailment"]
-        if not_entailment_score > self._threshold:
-            logger.warning(f"Detected refutation in the output: {prediction}")
+        entailment_score = prediction["entailment"]
+        if entailment_score < self._minimum_score:
+            logger.warning(f"Entailment score is below the threshold: {prediction}")
 
-            return output, False, not_entailment_score
+            return output, False, prediction["not_entailment"]
 
-        logger.debug(f"Not refutation in the output: {prediction}")
+        logger.debug(f"The output is factually consistent: {prediction}")
 
         return output, True, 0.0